Creating AWS S3 upload functionality.
