+++ .classpath
@@ -3,7 +3,6 @@
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry kind="src" path="src"/>
-	<classpathentry exported="true" kind="lib" path="lib/jcommon-1.0.16.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/jcommon-1.0.17.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/jfreechart-1.0.14-swt.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/jfreechart-1.0.14.jar"/>
@@ -13,5 +12,15 @@
 	<classpathentry exported="true" kind="lib" path="lib/eclipse-collections-10.0.0.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/eclipse-collections-api-10.0.0.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/j2html-1.4.0.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/aws-java-sdk-bundle-1.11.264.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/commons-codec-1.11.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/commons-logging-1.1.3.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/httpclient-4.5.9.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/httpcore-4.4.11.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jackson-annotations-2.6.0.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jackson-core-2.6.7.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jackson-databind-2.6.7.3.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jcommon-1.0.16.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/joda-time-2.8.1.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
