+++ META-INF/MANIFEST.MF
@@ -2,7 +2,7 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: %Bundle-Name
 Bundle-SymbolicName: O3SMeasures;singleton:=true
-Bundle-Version: 2.0.0
+Bundle-Version: 2.1.0
 Bundle-Vendor: %Bundle-Vendor
 Require-Bundle: org.eclipse.ui;bundle-version="3.109.0";visibility:=reexport,
  org.eclipse.core.resources;bundle-version="3.12.0";visibility:=reexport,
@@ -543,8 +543,7 @@ Export-Package: com.keypoint,
    org.jsefa.common.mapping",
  org.jsefa.xml.mapping.support;uses:="org.jsefa.xml.namespace",
  org.jsefa.xml.namespace
-Bundle-ClassPath: lib/jcommon-1.0.16.jar,
- lib/jcommon-1.0.17.jar,
+Bundle-ClassPath: lib/jcommon-1.0.17.jar,
  lib/jfreechart-1.0.14-swt.jar,
  lib/jfreechart-1.0.14.jar,
  .,
@@ -553,5 +552,15 @@ Bundle-ClassPath: lib/jcommon-1.0.16.jar,
  lib/log4j-1.2.17.jar,
  lib/eclipse-collections-10.0.0.jar,
  lib/eclipse-collections-api-10.0.0.jar,
- lib/j2html-1.4.0.jar
+ lib/j2html-1.4.0.jar,
+ lib/aws-java-sdk-bundle-1.11.264.jar,
+ lib/commons-codec-1.11.jar,
+ lib/commons-logging-1.1.3.jar,
+ lib/httpclient-4.5.9.jar,
+ lib/httpcore-4.4.11.jar,
+ lib/jackson-annotations-2.6.0.jar,
+ lib/jackson-core-2.6.7.jar,
+ lib/jackson-databind-2.6.7.3.jar,
+ lib/jcommon-1.0.16.jar,
+ lib/joda-time-2.8.1.jar
 Automatic-Module-Name: O3SMeasures
