+++ build.properties
@@ -6,7 +6,6 @@ bin.includes = plugin.xml,\
                lib/,\
                lib/jfreechart-1.0.14-swt.jar,\
                lib/jfreechart-1.0.14.jar,\
-               lib/jcommon-1.0.16.jar,\
                lib/jcommon-1.0.17.jar,\
                lib/log4j-1.2.17.jar,\
                icons/,\
@@ -19,10 +18,18 @@ bin.includes = plugin.xml,\
                .classpath,\
                lib/eclipse-collections-10.0.0.jar,\
                lib/eclipse-collections-api-10.0.0.jar,\
-               lib/j2html-1.4.0.jar
-jars.compile.order = .
+               lib/j2html-1.4.0.jar,\
+               lib/aws-java-sdk-bundle-1.11.264.jar,\
+               lib/commons-codec-1.11.jar,\
+               lib/commons-logging-1.1.3.jar,\
+               lib/httpclient-4.5.9.jar,\
+               lib/httpcore-4.4.11.jar,\
+               lib/jackson-annotations-2.6.0.jar,\
+               lib/jackson-core-2.6.7.jar,\
+               lib/jackson-databind-2.6.7.3.jar,\
+               lib/joda-time-2.8.1.jar,\
+               lib/jcommon-1.0.16.jar
 jars.extra.classpath = lib/commons-csv-1.0.jar,\
-                       lib/jcommon-1.0.16.jar,\
                        lib/jcommon-1.0.17.jar,\
                        lib/jfreechart-1.0.14-swt.jar,\
                        lib/jfreechart-1.0.14.jar,\
@@ -30,5 +37,14 @@ jars.extra.classpath = lib/commons-csv-1.0.jar,\
                        lib/log4j-1.2.17.jar,\
                        lib/eclipse-collections-10.0.0.jar,\
                        lib/eclipse-collections-api-10.0.0.jar,\
-                       lib/j2html-1.4.0.jar
+                       lib/j2html-1.4.0.jar,\
+                       lib/aws-java-sdk-bundle-1.11.264.jar,\
+                       lib/commons-codec-1.11.jar,\
+                       lib/commons-logging-1.1.3.jar,\
+                       lib/httpclient-4.5.9.jar,\
+                       lib/httpcore-4.4.11.jar,\
+                       lib/jackson-annotations-2.6.0.jar,\
+                       lib/jackson-core-2.6.7.jar,\
+                       lib/jackson-databind-2.6.7.3.jar,\
+                       lib/joda-time-2.8.1.jar
 src.includes = bin/
