+++ src/io/github/mariazevedo88/o3smeasures/javamodel/LackCohesionMethodsJavaModel.java
@@ -167,6 +167,7 @@ public class LackCohesionMethodsJavaModel implements IJavaModel<ICompilationUnit
 	 * 
 	 * @param methods
 	 */
+	@SuppressWarnings("deprecation")
 	private void checkMethodsWithSharedAttributes(IMethod[] methods){
 		IScanner scanner = null;
 		for (IMethod method : methods) {
