+++ src/io/github/mariazevedo88/o3smeasures/preferences/O3SMeasuresPreferencePage.java
@@ -2,6 +2,7 @@ package io.github.mariazevedo88.o3smeasures.preferences;
 
 import org.eclipse.jface.preference.DirectoryFieldEditor;
 import org.eclipse.jface.preference.FieldEditorPreferencePage;
+import org.eclipse.jface.preference.StringFieldEditor;
 import org.eclipse.ui.IWorkbench;
 import org.eclipse.ui.IWorkbenchPreferencePage;
 
@@ -25,7 +26,7 @@ public class O3SMeasuresPreferencePage extends FieldEditorPreferencePage impleme
 	@Override
 	public void init(IWorkbench workbench) {
 		setPreferenceStore(Activator.getDefault().getPreferenceStore());
-	    setDescription("O3SMeasures Preferences");
+		setTitle("O3SMeasures Preferences");
 	}
 
 	/**
@@ -33,7 +34,13 @@ public class O3SMeasuresPreferencePage extends FieldEditorPreferencePage impleme
 	 */
 	@Override
 	protected void createFieldEditors() {
+		
 		addField(new DirectoryFieldEditor("PATH", "&Directory preference:", getFieldEditorParent()));
+		
+		addField(new StringFieldEditor("BUCKET", "&AWS Bucket Name:", getFieldEditorParent()));
+		addField(new StringFieldEditor("REGION", "&AWS Bucket Region:", getFieldEditorParent()));
+		addField(new StringFieldEditor("ACCESSKEY", "&AWS Access Key:", getFieldEditorParent()));
+		addField(new StringFieldEditor("SECRETKEY", "&AWS Secret Key:", getFieldEditorParent()));
 	}
 
 }
