+++ src/io/github/mariazevedo88/o3smeasures/test/CyclomaticComplexityTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.TestInstance.Lifecycle;
 import io.github.mariazevedo88.o3smeasures.builder.MethodBuilder;
 import io.github.mariazevedo88.o3smeasures.measures.main.CyclomaticComplexity;
 import io.github.mariazevedo88.o3smeasures.structures.ItemMeasured;
-import io.github.mariazevedo88.o3smeasures.util.JavaParser;
+import io.github.mariazevedo88.o3smeasures.util.parsers.JavaParser;
 
 /**
  * A class test that executes CC measure test calculation
