+++ src/io/github/mariazevedo88/o3smeasures/test/FanOutTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
 
 import io.github.mariazevedo88.o3smeasures.measures.main.FanOut;
-import io.github.mariazevedo88.o3smeasures.util.JavaParser;
+import io.github.mariazevedo88.o3smeasures.util.parsers.JavaParser;
 
 /**
  * A class test that executes FOUT measure test calculation
