+++ src/io/github/mariazevedo88/o3smeasures/test/WeightMethodsPerClassTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
 
 import io.github.mariazevedo88.o3smeasures.measures.main.WeightMethodsPerClass;
-import io.github.mariazevedo88.o3smeasures.util.JavaParser;
+import io.github.mariazevedo88.o3smeasures.util.parsers.JavaParser;
 
 /**
  * A class test that executes WMC measure test calculation
