+++ src/io/github/mariazevedo88/o3smeasures/util/FileUpload.java
@@ -1,0 +1,21 @@
+package io.github.mariazevedo88.o3smeasures.util;
+
+/**
+ * Interface to implements file upload feature
+ * 
+ * @author Mariana Azevedo
+ * @since 25/01/2020
+ */
+public interface FileUpload {
+	
+	/**
+	 * Method to upload a file indicated in the method paramether
+	 * 
+	 * @author Mariana Azevedo
+	 * @since 25/01/2020
+	 * 
+	 * @param fileName
+	 */
+	public void upload(String fileName);
+
+}
